requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.2
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
